transaksis dan details

+++ database/migrations/2025_04_24_234406_create_transaksis_table.php
@@ -14,9 +14,8 @@ return new class extends Migration
         Schema::create('transaksis', function (Blueprint $table) {
             $table->id();
             $table->date('tanggal');
-            $table->decimal('total', 10, 2);
+            $table->integer('total');
             $table->timestamps();
-
         });
     }
